requirements.txt
streamlit
pandas
matplotlib
langchain
python-dotenv
openai
langchain-experimental